perf(sa): use u32 for suffix arrays (P0-4)
Reduce SA memory usage by 50% by storing positions as u32 instead of i64.
u32 supports sequences up to 4.3B bases (covers human genome).

- Added runtime check to reject sequences > u32::MAX with clear error
- TODO: Add feature flag for u64 SA for very large genomes

Files changed: sa.rs, search.rs, parallel_sa.rs

+++ src/search.rs
@@ -589,7 +589,7 @@ impl<'a> SaIndex<'a> {
     #[allow(clippy::too_many_arguments)]
     fn search_sa_simple(
         &self,
-        sa: &[i64],
+        sa: &[u32],
         text: &[u8],
         seed: &[u8],
         pairing: SeedPairing,
@@ -667,7 +667,7 @@ impl<'a> SaIndex<'a> {
     /// Helper to find range in SA for a specific character at offset
     fn get_sa_interval(
         &self,
-        sa: &[i64],
+        sa: &[u32],
         text: &[u8],
         start: usize,
         end: usize,
